Add initial tests file and remove unused ones
