+++ lambda/tox.ini
@@ -9,4 +9,4 @@ setenv =
     APP_CLIENT_ID = test_app_client_id
     AWS_DEFAULT_REGION = test_region
     
-commands = python -m unittest discover -p "test_*.py"
+commands = python3 -m unittest discover -p "test_verify.py"
